+++ estampes/tools/spec.py
@@ -1,0 +1,211 @@
+"""Toolbox for spectra
+
+Module providing tools to manipulate data relative to spectra.
+
+Methods
+-------
+broaden
+    Broadens a stick spectra with a given broadening function.
+convert_y
+    Returns parameters for the broadening to convert Y units.
+"""
+
+from math import ceil
+import typing as tp
+
+from estampes.tools.math import f_gauss, f_lorentz
+
+
+# ==============
+# Module Methods
+# ==============
+
+def broaden(xval: tp.Sequence[float],
+            yval: tp.Sequence[float],
+            xaxis: tp.Sequence[float],
+            funcname: str,
+            hwhm: float,
+            yfactor: float = 1.0,
+            xpower: int = 1,
+            ynorm: bool = False,
+            truncate: bool = False) -> tp.List[float]:
+    """Broadens a stick spectra with a given broadening function.
+
+    Performs the broadening operation given a list of transitions stored in
+      two separate arrays, `xval` and `yval`.
+    Returns the broadened spectrum for each point in `xaxis`.
+    `yaxis` is computed as:
+    yax[i] = sum_j yf*yv[j]*xax[i]**xpow*f(xax[i]-xv[j])
+    * yax: `yaxis`
+    * jf: `yfactor`
+    * yv: `yval`
+    * xax: `xaxis`
+    * xpow: `xpower`
+    * f: `function`
+    * xv: `xval`
+
+    Parameters
+    ----------
+    xval
+        List of X values.
+    yval
+        List of Y values.
+    xaxis
+        X axis of the broadened spectrum.
+    funcname
+        Broadening function: gaussian, lorentzian.
+    hwhm
+        Half-width at half-maximum (in unit of X).
+    yfactor
+        Scaling factor for the Y axis.
+    xpower
+        Power of X to include to the final Y value.
+    ynorm
+        Normalize Y (in this case, `yfactor` is ignored).
+    truncate
+        Truncate the broadening (for speed up).
+        Note: `xaxis` must be linear and regular (delta x constant).
+
+    Returns
+    -------
+    list
+        Y axis with the broadened bands.
+
+    Raises
+    ------
+    IndexError
+        `xaxis` is empty.
+    """
+    supported_func = {'gaussian': f_gauss, 'lorentzian': f_lorentz}
+    # Transform broadening function to be elemental
+    func = supported_func[funcname]
+    npoints = len(xaxis)
+    if npoints == 0:
+        raise IndexError('New X axis is empty')
+    yaxis = [0.0 for _ in npoints]
+    if not truncate:
+        for i in range(npoints):
+            for x0, y0 in zip(xval, yval):
+                yaxis[i] += func(xaxis[i], hwhm, x0, y0*yfactor)
+    else:
+        # hwtimes: Function is computer over an interval of 2*hwhm*hwtimes
+        # 3.5*hwhm represents 99.7% of the Gaussian area
+        # For lorentzian, the same coverage requires 212*hwhm, truncated to 15.
+        hwtimes = {'gaussian': 3.5, 'lorentzian': 15}[func]
+        xmin = xaxis[0]
+        dx = abs(xaxis[-1] - xaxis[0])/(npoints-1)
+        nhwpoints = int(ceil(hwhm*hwtimes/dx))
+        if xmin > xaxis[-1]:
+            # X axis is reversed
+            xmin = xaxis[-1]
+            for x0, y0 in zip(xval, yval):
+                i0 = -int((x0 - xmin)/dx) - 1
+                imin = max(i0 - nhwpoints, -npoints)
+                imax = min(i0 + nhwpoints, -1)
+                for i in range(imin, imax+1):
+                    yaxis[i] += func(xaxis[i], hwhm, x0, y0*yfactor)
+        else:
+            for x0, y0 in zip(xval, yval):
+                i0 = int((x0 - xmin)/dx)
+                imin = max(i0 - nhwpoints, 0)
+                imax = min(i0 + nhwpoints, npoints-1)
+                for i in range(imin, imax+1):
+                    yaxis[i] += func(xaxis[i], hwhm, x0, y0*yfactor)
+    if xpower > 0:
+        for i in range(npoints):
+            yaxis[i] *= xaxis[i]**xpower
+    if ynorm:
+        ymax = max(yaxis, key=abs)
+        for i in range(npoints):
+            yaxis /= ymax
+    return yaxis
+
+
+def convert_y(specabbr: str,
+              unit_to: str,
+              unit_from: str) -> tp.Tuple[float, int]:
+    """Returns parameters for the conversion of Y units.
+
+    Returns the scaling factor and the power of X needed to convert the
+      Y unit from `unit_from` to `unit_to`.
+    The units are strings composed of two parts, separated by colon:
+    - a description: DS, RS, RA, I, II...
+    - a unit: km/mol, cgs...
+    A default unit may be available, in which case, the unit can be
+      provided as:
+    `desc:`
+
+    Parameters
+    ----------
+    specabbr
+        Spectroscopy name (abbreviated version).
+    unit_to
+        Final Y unit.
+    unit_from
+        Original Y unit.
+
+    Returns
+    -------
+    float
+        Scaling factor on the Y unit.
+    int
+        Power of X to include to the final Y value.
+
+    Raises
+    ------
+    IndexError
+        Unrecognized spectroscopy or unit.
+    NotImplementedError
+        Spectroscopy or unit not yet available.
+
+    Notes
+    -----
+    * Recognized unit descriptions:
+        * `I` - intensity
+        * `II` - integrated intensity
+        * `DS` - dipole strength
+        * `RS` - rotatory strength
+    * Units
+        * `M-1` or `/M` - dm^3/mol
+        * `M-1cm-1` or `/M/cm` - dm^3/mol/cm
+    """
+    msgNYI = f'For {specabbr}: conversion from "{unit_from}" to "{unit_to}"' +\
+        ' not yet implemented.'
+    msgNA = f'Conversion from "{unit_from}" to "{unit_to}" not recognized ' +\
+        f'for {specabbr}'
+    # Initial setup
+    _spec = specabbr.upper()
+    res = unit_to.split(':')
+    if len(res) == 1:
+        _dest_unit = ''
+    else:
+        _dest_unit = res[1].lower().replace('.', '')
+    _dest_type = res[0].upper()
+    res = unit_from.split(':')
+    if len(res) == 1:
+        _src_unit = ''
+    else:
+        _src_unit = res[1].lower().replace('.', '')
+    _src_type = res[0].upper()
+    yfactor = 1.0
+    xunit = 0
+    # Test
+    if _spec == 'OPA':
+        if _dest_type == 'I':
+            if _dest_unit in ('', 'm-1l-1', '/m/cm'):
+                if _src_type == 'II':
+                    if _src_unit in ('m-1', '/m'):
+                        yfactor = 1.0
+                        xunit = 0
+                    else:
+                        raise NotImplementedError(msgNYI)
+                else:
+                    raise NotImplementedError(msgNYI)
+            else:
+                raise IndexError(msgNA)
+        else:
+            raise IndexError(msgNA)
+    else:
+        raise NotImplementedError(msgNYI)
+
+    return yfactor, xunit